% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{getDensities}
\alias{getDensities}
\title{Estimate densities
Uses greedy plaid covering to estimate per-batch global densities}
\usage{
getDensities(
  sce,
  dimred,
  group,
  ndims = NULL,
  k = NULL,
  alpha = 0.1,
  verbose = FALSE
)
}
\arguments{
\item{sce}{a `SingleCellExperiment` object}

\item{dimred}{character, the name of the reduced dimension slot in `sce`}

\item{group}{character, name in the `colData` slot of `sce` containing information
on the group/batch identity.}

\item{ndims}{numeric, the number of dimensions to consider in `dimred`. The
first `ndims` will be considered (1:ndims). Default is NULL meaning all
available dimensions will be used.}

\item{k}{numeric, the target number of hypercubes/boxes to use for plaid covering}

\item{alpha}{numeric, the tolerance on the final number of boxes given as
fraction of k}

\item{verbose}{logical, should the function print messages on the progress?
Default is `FALSE`.}
}
\value{
a vector containing the density for each level of `group`.
}
\description{
Estimate densities
Uses greedy plaid covering to estimate per-batch global densities
}
