% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{getIntervals}
\alias{getIntervals}
\title{Get intervals
Find which cells lie within a given set of non-contiguous intervals}
\usage{
getIntervals(points, d, l)
}
\arguments{
\item{points}{matrix, a cell x coordinates matrix of cell embeddings}

\item{d}{numeric, the dimension (column number of `points`)}

\item{l}{numeric, the size of the interval}
}
\value{
a data.frame where each cell is assigned to a specific interval number
}
\description{
Get intervals
Find which cells lie within a given set of non-contiguous intervals
}
