% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{getJointDensities}
\alias{getJointDensities}
\title{Estimate joint densities
Uses greedy plaid covering to estimate pairwise per-batch global densities}
\usage{
getJointDensities(
  sce,
  dimred,
  group,
  combs = NULL,
  ndims = NULL,
  k = NULL,
  alpha = 0.1,
  verbose = FALSE
)
}
\arguments{
\item{sce}{a `SingleCellExperiment` object}

\item{dimred}{character, the name of the reduced dimension slot in `sce`}

\item{group}{character, name in the `colData` slot of `sce` containing information
on the group/batch identity.}

\item{combs}{a data.frame containing pairwise combinations of group levels.
default is NULL meaning combinations will be constructed for all levels.}

\item{ndims}{numeric, the number of dimensions to consider in `dimred`. The
first `ndims` will be considered (1:ndims). Default is NULL meaning all
available dimensions will be used.}

\item{k}{numeric, the target number of hypercubes/boxes to use for plaid covering}

\item{alpha}{numeric, the tolerance on the final number of boxes given as
fraction of k}

\item{verbose}{logical, should the function print messages on the progress?
Default is `FALSE`.}
}
\value{
a named vector containing the density for each pair of `group` levels.
}
\description{
Estimate joint densities
Uses greedy plaid covering to estimate pairwise per-batch global densities
}
