% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDB.R
\name{.pHD}
\alias{.pHD}
\title{Robust partial Hausdorff distance from distance matrix}
\usage{
.pHD(A, B, q)
}
\arguments{
\item{A}{a matrix with the target set of points}

\item{B}{a matrix with the query set of points}

\item{q}{numeric, the rank of the minimum distance for robust calculations}
}
\value{
a numeric with the Hausdorff distance from A to B
}
\description{
Returns the  robust partial Hausdorff distance between two sets of points
}
